move stats card item to discrete file
